Main dynamically selects searchers
Main loads list of strings that provide searcher coordinates.
Later we can extract the strings into a file in a config directory
that is loaded, which will mean people can install plugins via a pip
install and update the config. We can eventually let folks update the
config using the app itself.

Main now delegates to searcher classes to determine if they are applicable
to the platform.