Maintain bindings on read
Before this everytime we tried to see what plugins we had we reloaded all of them.  Side effect of this is the objects returned from the first and second calls were different memory wise though value wise identical. No need for this. Caching calls to instantiate function so long as package list does not change.